Solved issue #1
I have written a simple Bash script that can be used on a macOS system to extract the SSID (Wi-Fi username) of the currently connected Wi-Fi network. This script utilises the airport utility that is included with macOS but is located within a framework, so its path is a bit lengthy.